Extract study from quizlet file
